P of a PID for the claw positions
Robot: added in teleopPeriodic the settings for pMotor.

robotContainer: added variable for pidClaw

Claw: eliminated the servo setting that wont be used anymore.

PIDClaw: created coding for the use of pid for the claw.

Teleop: erased the servo settings and added the joy for the PID claw (missing setting on TeleOp) and erased toggle for servo claw.

Constants: eliminated extra line 47, added constant for P_CLAW, and constants of kP, kI, and kD.